services/crawler/requirements.txt
httpx==0.27.0
beautifulsoup4==4.12.3
markdownify==0.12.1
pdfplumber==0.11.5
python-docx==1.1.2
python-pptx==1.0.2
PyYAML==6.0.1
tiktoken==0.6.0
playwright>=1.57.0,<2.0
openpyxl==3.1.5
